add experimental ytmusic playback for external songs
