+++ src/main/features/core/index.ts
@@ -4,3 +4,4 @@ import './player';
 import './remote';
 import './settings';
 import './discord-rpc';
+import './youtube';
