+++ src/renderer/features/musicbrainz/api/youtube-api.ts
@@ -1,0 +1,17 @@
+import { queryOptions } from '@tanstack/react-query';
+
+async function searchYoutube(query: string): Promise<Array<{ type: string; videoId?: string }>> {
+    if (typeof window !== 'undefined' && window.api?.youtube) {
+        return window.api.youtube.search(query);
+    }
+    return [];
+}
+
+export const youtubeQueries = {
+    search: (args: { query: string }) => {
+        return queryOptions({
+            queryFn: () => searchYoutube(args.query),
+            queryKey: ['youtube', 'search', args.query],
+        });
+    },
+};
